Major update
- Show syscalls
- Show threads even if thread don't have a BROFILER_START_THREAD
- Much better support for fibers
- Filters now work for whole timeline (show filter time for every frame)
- Much better visualisation for sync/work intervals
- Show context switch detailed info (even if switched to another
process)
- Fixed some crashes, when target application is x86

+++ BrofilerCore/CallstackCollector.h
@@ -15,13 +15,15 @@ namespace Brofiler
 	{
 		uint64 threadID;
 		uint64 timestamp;
-		size_t* callstack;
+		uint64* callstack;
 		uint8 count;
 	};
+
+
 	//////////////////////////////////////////////////////////////////////////
 	class CallstackCollector
 	{
-		// Packed callstack list: {ThreadID, Timestamp, Count, u64[Count]}
+		// Packed callstack list: {ThreadID, Timestamp, Count, Callstack[Count]}
 		typedef MemoryPool<uint64, 1024 * 32> CallstacksPool;
 		CallstacksPool callstacksPool;
 	public:
